requirements.txt
# Dépendances pour le testeur MTCaptcha
selenium>=4.15.0
requests>=2.31.0
python-dotenv>=1.0.0
webdriver-manager>=4.0.0
schedule>=1.2.0